requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.8.5
supabase>=2.6.0
